requirements.txt
python-dotenv~=1.0.1
requests~=2.31.0
tqdm~=4.66.2
customtkinter~=5.2.2
pillow~=10.3.0